some refactor.
