+++ src/main/scala/chiseltest/backends/VPIInterface.scala
@@ -5,8 +5,9 @@ import java.io.File
 import java.nio.channels.FileChannel
 
 import chisel3._
-import chiseltest.backends.VerilatorBackend
-import firrtl.ir.{Circuit, GroundType, IntWidth, Module, Port}
+import chiseltest.backends.{SimulatorInterface, VPIBackend}
+import firrtl.AnnotationSeq
+import firrtl.ir.{GroundType, IntWidth, Port}
 import logger.LazyLogging
 
 import scala.collection.immutable.ListMap
@@ -17,20 +18,11 @@ import scala.concurrent.{Await, ExecutionContext, Future, blocking}
 import scala.language.implicitConversions
 import scala.sys.process.{Process, ProcessLogger}
 
-/** Provides an interface layer to an external verilog simulation program through IPC channels
-  * Currently supported external programs are those generated by verilator and VCS.
-  *
-  * @param cmd The command to run as a Seq of strings
-  */
-private[chiseltest] class SimApiInterface(circuit: Circuit, cmd: Seq[String])
-  extends LazyLogging {
-  def getPorts(circuit: Circuit) = circuit.modules.collectFirst { case Module(_, name, ports, _) if name == circuit.main => ports }.get.partition(_.direction == firrtl.ir.Input)
-  //
-  // Construct maps for the input and output
-  // Remove zero length fields during the process
-  //
+class VPIInterface(val annotations: AnnotationSeq)
+  extends SimulatorInterface
+  with LazyLogging {
   val (inputsNameToChunkSizeMap, outputsNameToChunkSizeMap) = {
-    val (inputs, outputs) = getPorts(circuit)
+    val (inputs, outputs) = getTopPorts(annotations).partition(_.direction == firrtl.ir.Input)
 
     /** genChunk only will gen vaild Chunk now. */
     def genChunk(args: Port): Option[(String, Int)] = {
@@ -42,7 +34,7 @@ private[chiseltest] class SimApiInterface(circuit: Circuit, cmd: Seq[String])
           } else if (name.endsWith("clock")) {
             "clock"
           } else {
-            s"${circuit.main}.$name"
+            s"${getTopName(annotations)}.$name"
           }
           Some(n -> ((width.asInstanceOf[IntWidth].width.toInt - 1) / 64 + 1))
         case _ =>
@@ -69,9 +61,11 @@ private[chiseltest] class SimApiInterface(circuit: Circuit, cmd: Seq[String])
   private val _signalMap = mutable.HashMap[String, Int]()
   private val _chunks = mutable.HashMap[String, Int]()
   private val _logs = mutable.ArrayBuffer[String]()
+  private var _currentClock = 0
 
-  //initialize simulator process
-  private[chiseltest] val process = TesterProcess(cmd, _logs)
+
+  /** initialize simulator process*/
+  private[chiseltest] val process = TesterProcess(getCommand(annotations).get, _logs)
 
   // Set up a Future to wait for (and signal) the test process exit.
 
@@ -123,7 +117,6 @@ private[chiseltest] class SimApiInterface(circuit: Circuit, cmd: Seq[String])
   }
 
   private def throwExceptionIfDead(exitValue: Future[Int]) {
-    //    implicit val logger = new TestErrorLog
     if (exitValue.isCompleted) {
       val exitCode = Await.result(exitValue, Duration(-1, SECONDS))
       // We assume the error string is the last log entry.
@@ -256,10 +249,12 @@ private[chiseltest] class SimApiInterface(circuit: Circuit, cmd: Seq[String])
     isStale = false
   }
 
+
   private def takeStep(): Unit = {
     mwhile(!sendCmd(SIM_CMD.STEP)) {}
     mwhile(!sendInputs) {}
     mwhile(!recvOutputs) {}
+    _currentClock += 1
     dumpLogs()
   }
 
@@ -312,18 +307,15 @@ private[chiseltest] class SimApiInterface(circuit: Circuit, cmd: Seq[String])
     }
   }
 
-  private def start(): Unit = {
-    println(s"""STARTING ${cmd mkString " "}""")
-    mwhile(!recvOutputs) {}
+  /** interface to get current clock. */
+  def cycleCount: BigInt = _currentClock
 
-    //TODO: Figure out what the behavior of the interface on start up is
-    //    for (_ <- 0 until 5) {
-    //      mwhile(!sendCmd(SIM_CMD.RESET)) { }
-    //      mwhile(!recvOutputs) { }
-    //    }
+  private def start(): Unit = {
+    println(s"""STARTING ${getCommand(annotations) mkString " "}""")
+    mwhile(!recvOutputs) {}
   }
 
-  def poke(signal: String, value: BigInt) {
+  def poke(signal: String, value: BigInt): Unit = {
     if (inputsNameToChunkSizeMap contains signal) {
       _pokeMap(signal) = value
       isStale = true
@@ -361,7 +353,7 @@ private[chiseltest] class SimApiInterface(circuit: Circuit, cmd: Seq[String])
     }
   }
 
-  def step(n: Int) {
+  def step(n: Int): Unit = {
     update()
     (0 until n) foreach (_ => takeStep())
   }
@@ -450,16 +442,6 @@ private[chiseltest] object TesterProcess {
     // This makes everything written to stderr get added as lines to logs
     val processLogger = ProcessLogger(println, logs += _) // don't log stdout
     processBuilder run processLogger
-  }
-
-  //TODO: must figure this out before PR
-  def kill(sim: SimApiInterface) {
-    while (!sim.exitValue.isCompleted) sim.process.destroy
-    println("Exit Code: %d".format(sim.process.exitValue))
-  }
-
-  def kill(p: VerilatorBackend[_]) {
-    kill(p.simApiInterface)
   }
 }
